requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
tensorflow==2.20.0
numpy==2.2.6
pillow==12.1.0
sqlalchemy==2.0.46
mysql-connector-python==8.3.0
python-dotenv==1.2.1
pydantic==2.12.5
aiofiles==23.2.1
jinja2==3.1.6